+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-6602] Update td-cache, td-core and td-df-lib
+- [TD-6649] Add Business Concept file manager domain name
 
 ## [6.8.0] 2024-07-03